+++ extensions/quick-access/src/utils/common-utils.ts
@@ -1,15 +1,18 @@
-import { getPreferenceValues, getSelectedFinderItems, LocalStorage } from "@raycast/api";
+import { environment, getPreferenceValues, getSelectedFinderItems, LocalStorage } from "@raycast/api";
 import fse from "fs-extra";
-import { DirectoryInfo, DirectoryType } from "./directory-info";
+import { DirectoryInfo, DirectoryType, FileType } from "./directory-info";
 import { imgExt } from "./constants";
 import { parse } from "path";
 import Values = LocalStorage.Values;
 import { getFinderInsertLocation } from "./applescript-utils";
+import fileUrl from "file-url";
 
 export const commonPreferences = () => {
   const preferencesMap = new Map(Object.entries(getPreferenceValues<Values>()));
   return {
     autoCopyLatestFile: preferencesMap.get("autoCopyLatestFile") as boolean,
+    rememberTag: preferencesMap.get("rememberTag") as boolean,
+    primaryAction: preferencesMap.get("primaryAction") as string,
     fileShowNumber: preferencesMap.get("fileShowNumber") as string,
     sortBy: preferencesMap.get("sortBy") as string,
   };
@@ -80,6 +83,26 @@ export const isDirectoryOrFile = (path: string) => {
   return DirectoryType.FILE;
 };
 
+export const isDirectoryOrFileForFile = (path: string) => {
+  try {
+    const stat = fse.lstatSync(path);
+    const parsedPath = parse(path);
+    if (stat.isDirectory()) {
+      return FileType.FOLDER;
+    }
+    if (stat.isFile()) {
+      if (imgExt.includes(parsedPath.ext)) {
+        return FileType.IMAGE;
+      }
+      return FileType.FILE;
+    }
+  } catch (e) {
+    console.error(String(e));
+    return FileType.FILE;
+  }
+  return FileType.FILE;
+};
+
 export const checkDuplicatePath = (path: string, localDirectory: DirectoryInfo[]) => {
   //check duplicate
   let duplicatePath = false;
@@ -119,7 +142,7 @@ export const getDirectoryFiles = (directory: string) => {
         id: "files_" + (timeStamp + index),
         name: parsedPath.base,
         path: parsedPath.dir + "/" + parsedPath.base,
-        type: isDirectoryOrFile(value),
+        type: isDirectoryOrFileForFile(value),
         modifyTime: getModifyTime(value),
       };
     });
@@ -163,11 +186,71 @@ export const getFileShowNumber = (fileShowNumber: string) => {
       return 3;
     case "5":
       return 5;
-    case "8":
-      return 8;
     case "10":
       return 10;
     default:
       return -1;
   }
+};
+
+function getFileSize(size: number) {
+  if (!size) return "0 KB";
+  const num = 1024.0; //byte
+  if (size < num) return size + " B";
+  if (size < Math.pow(num, 2)) return (size / num).toFixed(2) + " KB"; //kb
+  if (size < Math.pow(num, 3)) return (size / Math.pow(num, 2)).toFixed(2) + " MB"; //M
+  if (size < Math.pow(num, 4)) return (size / Math.pow(num, 3)).toFixed(2) + " G"; //G
+  return (size / Math.pow(num, 4)).toFixed(2) + " T"; //T
+}
+
+const assetPath = environment.assetsPath;
+const raycastIsLightTheme = () => {
+  return environment.theme == "light";
+};
+
+//Detail content: Image
+export const getFileContent = (filePath: string) => {
+  let detailContent = "";
+  let preview = "";
+  let size = "";
+  const parsePath = parse(filePath);
+  try {
+    if (!isEmpty(filePath)) {
+      const fileStat = fse.statSync(filePath);
+      if (fileStat.isDirectory()) {
+        const files = fse.readdirSync(filePath);
+        const isNormalFile = files.filter((value) => !value.startsWith("."));
+        const previewIcon = raycastIsLightTheme() ? "folder-icon.png" : "folder-icon@dark.png";
+        preview = `![](${fileUrl(assetPath + "/" + previewIcon)})\n`;
+        size = `**Sub-files**: ${isNormalFile.length}`;
+      } else {
+        if (imgExt.includes(parsePath.ext)) {
+          preview = `![](${fileUrl(filePath)})\n`;
+        } else {
+          preview = `![](${fileUrl(assetPath + "/file-icon.png")})\n`;
+        }
+        size = `**Size**: ${getFileSize(fileStat.size)}`;
+      }
+
+      detailContent = `${preview}
+      
+------
+
+**Name**: ${parsePath.base}
+
+**Where**: ${parsePath.dir}
+
+${size}
+
+**Created**: ${new Date(fileStat.birthtime).toLocaleString()}
+
+**Modified**: ${new Date(fileStat.mtime).toLocaleString()}
+
+**Last opened**: ${new Date(fileStat.atime).toLocaleString()}
+`;
+    }
+  } catch (e) {
+    console.error(String(e));
+  }
+  return detailContent;
 };