+++ extensions/quick-access/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Quick Access Changelog
 
+## [Add Open With] - 2022-04-25
+
+- Add Detail view
+- Add new filter tag
+- Add actions **Open With...**
+- Add preferences **Primary action** : Set the primary action in Search Pinned Directories command
+- Update shortcuts for certain actions
+
 ## [Initial Version] - 2022-04-22
 
 - Quick copy, open, reveal the latest files in pinned Finder directories.